+++ android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.transportation_management_system_frontend",
+    package = "com.tms.customer",
 )
 
 android_resource(
     name = "res",
-    package = "com.transportation_management_system_frontend",
+    package = "com.tms.customer",
     res = "src/main/res",
 )
