+++ android/app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.transportation_management_system_frontend">
+  package="com.tms.customer">
 
     <uses-permission android:name="android.permission.INTERNET" />
