+++ android/app/src/main/java/com/tms/customer/MainActivity.java
@@ -1,4 +1,4 @@
-package com.transportation_management_system_frontend;
+package com.tms.customer;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "Transportation_Management_System_Frontend";
+    return "ShoberDelivery";
   }
 }
