+++ android/app/src/main/java/com/tms/customer/MainApplication.java
@@ -1,4 +1,4 @@
-package com.transportation_management_system_frontend;
+package com.tms.customer;
 
 import android.app.Application;
 import android.content.Context;
@@ -63,7 +63,7 @@ public class MainApplication extends Application implements ReactApplication {
          We use reflection here to pick up the class that initializes Flipper,
         since Flipper library is not available in release mode
         */
-        Class<?> aClass = Class.forName("com.transportation_management_system_frontend.ReactNativeFlipper");
+        Class<?> aClass = Class.forName("com.tms.customer.ReactNativeFlipper");
         aClass
             .getMethod("initializeFlipper", Context.class, ReactInstanceManager.class)
             .invoke(null, context, reactInstanceManager);
