+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Transportation_Management_System_Frontend</string>
+    <string name="app_name">Shober Delivery</string>
 </resources>
