+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "Transportation_Management_System_Frontend",
-  "displayName": "Transportation_Management_System_Frontend"
+  "name": "ShoberDelivery",
+  "displayName": "Shober Delivery"
 }
