+++ ios/Podfile
@@ -3,7 +3,7 @@ require_relative '../node_modules/@react-native-community/cli-platform-ios/nativ
 
 platform :ios, '11.0'
 
-target 'Transportation_Management_System_Frontend' do
+target 'ShoberDelivery' do
   config = use_native_modules!
 
   use_react_native!(
@@ -14,7 +14,7 @@ target 'Transportation_Management_System_Frontend' do
 
   pod 'RNVectorIcons', :path => '../node_modules/react-native-vector-icons'
 
-  target 'Transportation_Management_System_FrontendTests' do
+  target 'ShoberDeliveryTests' do
     inherit! :complete
     # Pods for testing
   end
