+++ ios/ShoberDeliveryTests/ShoberDeliveryTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface Transportation_Management_System_FrontendTests : XCTestCase
+@interface ShoberDeliveryTests : XCTestCase
 
 @end
 
-@implementation Transportation_Management_System_FrontendTests
+@implementation ShoberDeliveryTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
